gfs2_edit: Log descriptor continuation blocks print wrong info
Log descriptor continuation blocks are printing misleading
information. First, the offset from the start of the journal is
reported incorrectly. Second, it's trying to print the block
types when it shouldn't. The faulty output looks like this:
0x40e9 (j+4093000): Log descriptor continuation block
                    0x48202b none      0x481e2d none      0x481c2f none      0x481a31 none
In practice, these are used to continue the job of "revoke" log
descriptors, which have no blocks that follow, so it should not
try to report the types of the blocks that follow.

This patch fixes the journal offset calculation and skips the
reporting of the block types that follow, so the output looks
is given correctly as:
0x40e9 (j+4093): Log descriptor continuation block
         0x48202b        0x481e2d        0x481c2f        0x481a31