mkfs.gfs2: Always validate the locktable option
When mkfs.gfs2 is run with -p lock_nolock and the -t <locktable> option
is also used, validation of the locktable is skipped. Make sure the -t
option is always checked whether it's required (with lock_dlm/lock_gulm)
or not (with lock_nolock).

Also add some tests which ensure the locktable is validated in this
scenario.

Signed-off-by: Andrew Price <anprice@redhat.com>