fsck.gfs2: Add ability to fix rindex file size
Before this patch, fsck.gfs2 checked the size of rindex, and if
it wasn't evenly divisible by the size of an rindex entry, it
threw up its hands and gave up. This patch allows fsck.gfs2 to
repair the rindex size, basically rounding down to the largest
number of rgrps that will fit.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>