libgfs2: Support the new dirent de_cookie field
Detect whether the new de_cookie field (kernel commit 471f3db27) is
present in struct gfs2_dirent in the configure script and update the
metadata description to support that field.

It also seems that meta.c was not using the GFS2_HAS_* defines from
clusterautoconfig.h and that was causing a unit test to fail with

    gfs2_dirent: __pad: offset is 28, expected 26
    gfs2_dirent: size mismatch between struct 40 and fields 38

This commit adds an #include to fix that.

Fixes Rawhide bz#1307532

Signed-off-by: Andrew Price <anprice@redhat.com>