glocktop: Fix a tight loop under nohup
When glocktop is run under nohup, /dev/null is redirected to stdin,
which flags up EOF and causes the file descriptor to always be in a
ready state when select()ed. This causes select() to return immediately
instead of timing out after the refresh delay.

Remove stdin from the fd_set when we get EOF to avoid getting into a
tight loop when run under nohup.

Also fix up some messages relating to the refresh interval.

Signed-off-by: Andrew Price <anprice@redhat.com>