fsck.gfs2: Move pass5 to immediately follow pass1
The purpose of pass1 is to make sure all bits are represented
properly for inodes. After that, the bitmaps should be kept in
sync by subsequent passes. The purpose of the old pass5 was really
to free any blocks that were not linked somehow to dinodes.

This patch moves that reconciliation to immediately follow pass1.
That allows us to free the bitmap memory sooner, making subsequent
passes use the rgrp bitmaps for their work. The purpose is to
reduce the memory footprint of fsck.gfs2.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>