fsck.gfs2: Separate out functions that may only be done after pass1
This patch adds a new header file and source file which has the
purpose of separating out the functions that may only be called
after pass1 from the functions that pass1 may call. This is to
ensure that there will never be inconsistencies between blockmap
(which pass1 uses) and bitmap (which subsequent passes use).

Signed-off-by: Bob Peterson <rpeterso@redhat.com>