fsck.gfs2: pass1b shouldn't complain about non-bitmap blocks
One of the previous patches switched pass1b to stop using the
blockmap and start using the bitmap. One major difference between
them is that the blockmap has a value (0 - 3) for every block on
the device, whereas the bitmap has a value (0 - 3) for every block
in the file system. It does not have a value for blocks that are
on the device but aren't in the file system, such as the superblock
and the blocks preceding it, the blocks reserved for the rgrps
and their respective bitmaps.

That means the block_map function returned 0 for non-fs blocks,
but the bitmap function returned -1 (meaning invalid block).
The result was that pass1b complained about every non-fs block.

This patch makes pass1b check the return code from function
bitmap_type() and if it's a negative number, it skips over it.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>