fsck.gfs2: Use di_entries to determine if lost+found was created
Before this patch, function make_sure_lf_exists compared the bitmap
to the blockmap to determine if the lost+found directory was created.
Since we're phasing out the blockmap, and only want to use the
bitmap, we need an alternative. This patch changes it to check the
number of directory entries in the root directory has increased.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>