fsck.gfs2: adjust rgrp inode count when fixing bitmap
Before this patch, fsck.gfs2 would adjust the rgrp bitmaps and
rgrp free space when freeing a block. Later, in pass5, it would
fix the rgrp count of dinodes as needed. Now that pass5 is done
immediately after pass1, we need the other passes to make sure
to keep the dinode count in the rgrps in sync. Therefore, we
adjust it as we go.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>