fsck.gfs2: Fix GFS1 "used meta" accounting bug
Before this patch, fsck.gfs2 was grouping all data blocks and
non-inode metadata blocks in the same category. That's wrong
because the GFS1 "used meta" block count counts data blocks as
metadata blocks. This patch fixes it so that GFS1 "used metadata"
has its own count and the accounting is done properly.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>