fsck.gfs2: pass1b is too noisy wrt gfs1 non-dinode metadata
For GFS1 file systems, a block marked as metadata may either be a
dinode or any other kind of metadata. For GFS2 file systems, it makes
sense to complain when we find something that we think is a dinode,
but isn't. For GFS1 file systems, non-dinode metadata is a common
thing, so it makes sense to keep quiet about it.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>