fsck.gfs2: Fix rgrp dinode accounting bug
This patch fixes a bug whereby the count of dinodes in an rgrp can
get off because it's not adjusted properly when the bit is changed
from dinode to free. It's not adjusted because it's not found in
the inode tree or dirtree. The patch changes it so that the bitmap
is changed before deleting those things.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>