gfs2_edit: Make 'j' jump to journaled block on log descriptors
The jump command ordinarily jumps to the absolute block in the file
system. That makes sense for almost all cases. For log descriptors,
you're usually more interested in the journaled blocks than the
actual blocks. So this patch makes 'j' jump to the journaled version
of the block rather than the absolute block, when you're looking
at a log descriptor block.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>