fsck.gfs2: Speed up fsck.gfs2 with a new inode rgrp pointer
This patch tries to speed up fsck.gfs2 by adding a new field in the
in-core inode which points to the inode's rgrp. Most blocks will be
contained within the rgrp, so we save time looking up the rgrp by
first checking if the desired inode block falls within the inode's
rgrp.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>