fsck.gfs2: "undo" functions can stop too early on duplicates
This patch addresses a bug in which blocks may not be freed properly
because the "undo" processing in pass1 can stop early when duplicate
data blocks are encountered. What happens is this: The "undo" process
called from metawalk.c is coded to stop its undo when it encounters
the block containing corruption. However, if the block containing the
corruption is actually the second reference to the same block, inside
the same dinode, the undo process stops early when it encounters the
first reference. The problem is that it's coded to stop when it sees
a reference to the corrupt block. That's wrong. It should really
stop when it sees the reference to that block that triggered the
"unrecoverable" error. In other words, when it hits that exact
metadata block and its offset. This patch scraps the old system of
looking for the corrupt block in favor of looking for the corrupt
reference, including metadata block and offset within that block.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>