fsck.gfs2: link count checking wrong inode's formal inode number
This patch fixes a bug whereby inodes that aren't in the dirtree or
the inodetree are checking the wrong formal inode number. Function
incr_link_count checks the wrong inode value, and thus reports an
error where there are none. Also, when the error is reported back,
it reports the wrong value, using a copy of the inum value. In this
case, it should look up the dentry's inode's formal inode number.
This should only hurt performance in error cases.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

This is a RHEL7 port of this upstream patch:
https://git.fedorahosted.org/cgit/gfs2-utils.git/commit/?id=ad19203f04caa3fb1c777250a21fe1a968999e8f

rhbz#1350597