gfs2-utils: Build with -D_FORTIFY_SOURCE=2
A survey of popular Linux distros shows that they all compile their
packages with _FORTIFY_SOURCE=2 defined so it makes sense to enable it
upstream too, in order to fix any issues it highlights before they reach
the distros. It will also allow us to factor the overhead it adds into
perf measurements.

Briefly, using this flag adds buffer overflow detection to some
functions, and enables -Wunused-result warnings.

It requires the optimization level to be -O1 or higher so it is only
added to CPPFLAGS when --enable-debug is not used.

Signed-off-by: Andrew Price <anprice@redhat.com>