gfs2_edit: Fix unaligned accesses due to saved_metablock size
Each saved block is prefixed by a saved_metablock structure that holds
its block number and data length. Unfortunately this structure is 10
bytes in size, which means that storing the metadata header in memory
immediately following it causes accesses to the header fields to be
unaligned. This causes testsuite failures on sparc64. Instead, on
restore, use a separate buffer for the block data to ensure aligned
accesses to those fields.

Reported-By: Valentin Vidic <Valentin.Vidic@CARNet.hr>
Signed-off-by: Andrew Price <anprice@redhat.com>