fsck.gfs2: Repair rindex entries that were destroyed
If an rindex entry was destroyed, the rindex was skipped, which
means its rgrp never gets inserted into the rgrp tree. The trouble
is, that causes a off-by-one-entry problem with all subsequent
rindex entries. So if rindex entry R is destroyed, then rebuilt,
originally good entry R+1 is checked against new entry for R.

Function rg_repair already had the capability to insert new rindex
entries, but only at the end. This would be typical for a gfs2_grow
that failed and left a bunch of rgrps on the device, but the rindex
was never synced properly.

This patch allows the function to insert catch-up entries in the
middle, which re-syncs the actual versus expected rgrp trees.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>