gfs2_edit savemeta: Save corrupt rgrp headers
Previously, savemeta used gfs2_rgrp_read() to read resource groups. This
includes a metadata check which fails the read if the resource groups
are corrupt. Since we want to save corruption in order to analyze it, we
need to read the rgrp headers without checking them. Use a new,
simplified rgrp_read() function for that.

Also avoid the effect of a similar metadata check in
get_gfs_struct_info() when saving rgrps.

Signed-off-by: Andrew Price <anprice@redhat.com>