mkfs.gfs2: Warn when device is misaligned
Normally one layer of the I/O stack will adjust for a non-zero
alignment_offset value and expose it to upper layers as zero. If a
misalignment is not accounted for, mkfs.gfs2 will see a non-zero
alignment_offset and should report the misalignment to the user as it
could harm performance.

Signed-off-by: Andrew Price <anprice@redhat.com>