libgfs2: Issue one write per rgrp when creating them
Previously mkfs.gfs2 issued a write for each block of an rgrp. On
systems where the page size is much larger than the block size, this
incurs a read-modify-write overhead and can slow down mkfs.gfs2
considerably. Instead, allocate a single, aligned buffer for the bitmap
blocks and write them all in one go.

On a system with 64K pages and a 300TB block device, this speeds up
mkfs.gfs2 from over an hour to less than 3 minutes.

Signed-off-by: Andrew Price <anprice@redhat.com>