gfs2-utils: Add fscan
Add Steve Whitehouse's experimental fscan program/framework which is
intended to be the base for a possible new design of fsck.gfs2 that
tries to maximize i/o throughput from the outset.

Tweaks added to the original code to use libaio and the existing device
limits gathering code and byteswapping macros in libgfs2, along with
some general tidying.

Signed-off-by: Andrew Price <anprice@redhat.com>