gfs2_edit: Print offsets of indirect pointers
When printing indirect pointers gfs2_edit doesn't keep track of the
location of the pointer within the indirect block and only prints an
index for each non-zero pointer. Instead, store the offset of each
non-zero pointer and print it when displaying indirect blocks, as it is
done in interactive mode.

Resolves: rhbz#1518938

Signed-off-by: Andrew Price <anprice@redhat.com>