mkfs.gfs2: Improve alignment of first resource group
Currently the first rgrp is aligned to the whole stripe width and the
second rgrp is aligned to (stripe width + 1 stripe unit) and so on, to
spread them across an array. However, that means that there could be a
large amount of space wasted between the superblock and the first
resource group, and can result in the iovec used to zero that space
exceeding IOV_MAX and failing mkfs.gfs2 (since 6cefaf33d5) if the array
has a sufficiently large number of LUNs. Instead, align the first
resource group to a stripe unit so that the gap is minimised. Resource
groups are still spread across the array as the alignment of subsequent
ones are handled separately.

Resolves: rhbz#1698858

Signed-off-by: Andrew Price <anprice@redhat.com>