libgfs2: Speed up bread()
While bread() is still used in many hard-to-refactor places it should be
as performant as possible. Implementing it in terms of breadm() adds
extra iovec allocation overhead so simplify it by implementing it with a
simpler bget() and pread().

This only shows a 0.5% speed up in testing but that scales up to an hour
off a 300TB fsck.

Signed-off-by: Andrew Price <anprice@redhat.com>