fsck.gfs2: Retain context for indirect pointers in ->check_metalist
When the pass->check_metalist() functions are called, the pointer is
thrown away and the function is just called with a block address,
meaning that ->check_metalist() is unable to fix the pointer itself if
it deems the block pointed to is garbage.

Add an iptr structure which collects together the inode, the metadata
block buffer and the offset of the pointer in that buffer. The iptr is
now passed to ->check_metadata instead of the separate inode and block
address arguments.

Resolves: rhbz#1487726

Signed-off-by: Andrew Price <anprice@redhat.com>