restoremeta: Combine restore_init() and open_metadata()
Later we'll need to know which compression method to use for the
metadata file in order to initialise a valid metadata file descriptor,
and that requires reading an initial chunk. So it makes sense to combine
the open_metadata() function with restore_init() which reads the
preamble from the metadata file.

Signed-off-by: Andrew Price <anprice@redhat.com>