restoremeta: Add bzip2 support
Implement the required functions for adding bzip2 support and add
functions for reading the initial chunk so that we can try each
compression method. gzip has to be tried last as zlib's read function
doesn't report if the file is not in gzip format, it just returns the
data as-is. Adds a dependency on libbzip2.

Signed-off-by: Andrew Price <anprice@redhat.com>