libgfs2: Remove bi_bh from struct gfs2_bitmap
Replace bi_bh with a pointer to the start of the block within the
multi-block buffer containing the whole resource group. That way we can
allocate and read the whole resource group in one go and we're not
limited to single-block i/o's. The bh.b_modified flag has been pushed
down into the gfs2_bitmap as bi_modified. This could be obsoleted in
future using more deterministic writeback strategies.

Signed-off-by: Andrew Price <anprice@redhat.com>