savemeta: Don't save bad xattr blocks twice
This could only happen if the xattr block has a corrupt mh_type.

We've already saved the xattr block by the time we figure out that we
can't recognise its mh type so don't bother saving it again there.

Also relax mh type checking for the xattr blocks so that we save as much
of the (possibly corrupt) fs as possible. Just use the mh_type field to
decide whether to save indirect pointers or xattrs.

Signed-off-by: Andrew Price <anprice@redhat.com>