Reorganise unit tests
Move the libgfs2 unit tests out of the main testsuite and into the
libgfs2 directory. They are still run in 'make check' but the new
organisation means we have the tests closer to the code that they're
testing and the main testsuite script no longer needs exceptions for
running outside of the source tree (e.g. for CI).

Also rework the libgfs2 unit tests so that they're built as several
suites in one executable. This simplifies the build rules.

The setup and teardown code for the mocked structures in the rgrp.c unit
tests is now implemented as libcheck 'fixtures' that are run for each
test case.

The automake rules for building unit tests are now separated into a
checks.am file that is included depending on whether libcheck is
available.

Signed-off-by: Andrew Price <anprice@redhat.com>