mkfs.gfs2: make are_you_sure EOF happy
This patch adds a check for getline() error, as the manpage states out
that -1 is returned in EOF case we answer such case and possible
other errors with no. If you currently press ctrl-d during this question
an endless loop of asking this question over and over again will occur.
However this patch changes to handle errors with no and prints a newline
so the user have a new clean shell prompt afterwards. If getline reads
zero characters .e.g '\n', the question is if we are sure is asked again.

Signed-off-by: Alexander Aring <aahringo@redhat.com>
Signed-off-by: Andrew Price <anprice@redhat.com>